tmp-test: use the header key to verify PAC server signature
I think there a bug with PAC server signature, which only exhibits itself
now with failing tests because we now use PACs in the tests-suit.

These tests do a kinit then change the krbtgt key and then try to get
a service ticket with the old tgt, this causes the server signature
verification to fail (server signature is also krbtgt in the tgt
case), because the PAC does not convey the used kvno.

I think the same key used to encrypt the ticket should be used for
PAC server signature, like U2U tickets are signed witht the tgt
session-key and not with the longterm key, and so krbtgt should be
no different and the header key should be used.

An alternative would be to just try a couple of previous kvnos like
we do in MIT kdb_test plugin. We might need this anyway for the KDC
signature, especially in S4U2Proxy request.